Relize

+++ FourthProgramSimple/Function.h
@@ -8,12 +8,12 @@ vector<string> SubGenerator(const string& str);
 
 template <typename T = string>
 void print(const vector<T>& arr) {
-	for (int i = 0; i < arr.size(); ++i) {
+	for (size_t i = 0; i < arr.size(); ++i) {
 		cout << arr[i] << endl;
 	}
 }
 
-string IndexBuilder(int index, int subLength);
+string IndexBuilder(size_t index, size_t subLength);
 
 string stringReplace(const string& source, const string& toReplace, const string& replaceWith);
